move the mainnet/testnet option to advanced options
