+++ primer-android/res/layout/fragment_hot_option.xml
@@ -75,11 +75,6 @@
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"/>
 
-                <net.bither.ui.base.SettingSelectorView
-                    android:id="@+id/ssv_choose_net"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"/>
-
                 <LinearLayout
                     android:id="@+id/ll_switch_to_cold"
                     android:layout_width="match_parent"
