+++ primer-android/res/values-zh-rHK/strings.xml
@@ -442,7 +442,7 @@
     <string name="launch_sequence_switch_to_cold">切換到冷錢包</string>
     <string name="launch_sequence_switch_to_cold_warn">確認切換到冷錢包模式?\n\n冷錢包應該運行在離線的備用手機上.\n熱錢包可以監控冷錢包.\n一旦切換成冷錢包便不能切換回熱錢包.</string>
 
-    <string name="choose_net">選擇網絡（空錢包可以用）</string>
+    <string name="choose_net">選擇網絡</string>
     <string name="choose_mainnet">主網</string>
     <string name="choose_testnet">測試網</string>
     <string name="choose_net_warn">為保證網絡切換成功，建\n議8秒後重啟壹下錢包。</string>
