+++ primer-android/res/values-zh/strings.xml
@@ -442,7 +442,7 @@
     <string name="launch_sequence_switch_to_cold">切换到冷钱包</string>
     <string name="launch_sequence_switch_to_cold_warn">确认切换到冷钱包模式?\n\n冷钱包应该运行在离线的备用手机上.\n热钱包可以监控冷钱包.\n一旦切换成冷钱包便不能切换回热钱包.</string>
 
-    <string name="choose_net">选择网络（空钱包可以用）</string>
+    <string name="choose_net">选择网络</string>
     <string name="choose_mainnet">主网</string>
     <string name="choose_testnet">测试网</string>
     <string name="choose_net_warn">为保证网络切换成功，建\n议8秒后重启一下钱包。</string>
